requirements.txt
# Core libraries for data loading and processing
pandas~=2.2.3
numpy~=2.2.5
torch

# Hugging Face libraries for datasets and embeddings
datasets~=3.5.1
sentence-transformers~=4.1.0

faiss-cpu~=1.11.0
Flask~=3.1.1